// packages/codeboltjs/src/modules/debug.d.ts
import { DebugAddLogResponse, OpenDebugBrowserResponse } from '@codebolt/types/sdk';
export declare enum logType {
    info = "info",
    error = "error",
    warning = "warning"
}
export declare const debug: {
    /**
     * Sends a log message to the debug websocket and waits for a response.
     * @param {string} log - The log message to send.
     * @param {logType} type - The type of the log message (info, error, warning).
     * @returns {Promise<DebugAddLogResponse>} A promise that resolves with the response from the debug event.
     */
    debug: (log: string, type: logType) => Promise<DebugAddLogResponse>;
    /**
     * Requests to open a debug browser at the specified URL and port.
     * @param {string} url - The URL where the debug browser should be opened.
     * @param {number} port - The port on which the debug browser will listen.
     * @returns {Promise<OpenDebugBrowserResponse>} A promise that resolves with the response from the open debug browser event.
     */
    openDebugBrowser: (url: string, port: number) => Promise<OpenDebugBrowserResponse>;
};
export default debug;